% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmsunburst-package.r, R/ibmsunburst.R
\docType{package}
\name{ibmsunburst}
\alias{ibmsunburst}
\alias{ibmsunburst-package}
\title{ibmsunburst.}
\usage{
ibmsunburst(json_path, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{json_path}{Path to JSON data (IBM Watson Personality Insights service)}
}
\description{
Produce an IBM Sunburst Diagram from a JSON data output from
IBM Watson Personality Insights service
}
\examples{
json_path <- system.file("extdata", "profiles",
                         "en_v2.json", package = "ibmsunburst")

}
